Proyect content to kanban card
