+++ src/app/modules/roadmap/components/kanban-card/kanban-card.component.html
@@ -1,10 +1,8 @@
 <div class="kanban-card" [ngClass]="'kanban-card--' + color">
     <span class="kanban-card__status" [ngClass]="'kanban-card__status--' + color">Planned</span>
-    <h3 class="kanban-card__title">More comprehensive reports</h3>
+    <ng-content select="[kanban-card-title]"></ng-content>
     <div class="kanban-card__body">
-        <p class="kanban-card__text">
-            It would be great to see a more detailed breakdown of solutions.
-        </p>
+        <ng-content select="[kanban-card-text]"></ng-content>
         <ng-content></ng-content>
     </div>
 </div>
