--- src/app/modules/roadmap/components/kanban-card/kanban-card.component.ts
@@ -1,11 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-@Component({
-  selector: 'app-kanban-card',
-  templateUrl: './kanban-card.component.html',
-  styleUrls: ['./kanban-card.component.scss']
-})
-export class KanbanCardComponent  {
-  @Input() color: 'primary' | 'link' | 'tertiary' | 'none' = 'none';
-
-}
