+++ src/app/modules/roadmap/components/kanban-card/kanban-card.module.ts
@@ -1,14 +1,20 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { KanbanCardComponent } from './kanban-card.component';
+import { KanbanCardComponent, KanbanCardTextDirective, KanbanCardTitleDirective } from './kanban-card';
 
 @NgModule({
   declarations: [
-    KanbanCardComponent
+    KanbanCardComponent,
+    KanbanCardTitleDirective,
+    KanbanCardTextDirective
   ],
   imports: [
     CommonModule
   ],
-  exports: [KanbanCardComponent]
+  exports: [
+    KanbanCardComponent,  
+    KanbanCardTitleDirective,
+    KanbanCardTextDirective
+  ]
 })
 export class KanbanCardModule { }
