+++ src/app/modules/roadmap/layouts/roadmap-detail/roadmap-detail.component.html
@@ -11,5 +11,8 @@
             <ui-icon name="plus"></ui-icon> {{ t('createFeedbackLabelButton') }}
         </button>
     </ui-toolbar>
-   <app-kanban-card [color]="'link'"></app-kanban-card>
+   <app-kanban-card [color]="'link'">
+        <h3 kanban-card-title>More comprehensive reports</h3>
+       <p kanban-card-text>It would be great to see a more detailed breakdown of solutions.</p>
+   </app-kanban-card>
 </section>
